@@ -2,33 +2,38 @@ apply plugin: 'checkstyle'
 apply plugin: 'com.android.application'
 apply plugin: 'jacoco'
 apply plugin: 'realm-android'
+apply plugin: 'io.fabric'
 
 ext {
     androidTestVer = "0.5"
-    buildToolsVer = "26.0.0"
+    buildToolsVer = "27.0.3"
     butterknifeVer = "8.8.1"
     checkstyleVer = '7.7'
-    compileSdkVer = 25
+    compileSdkVer = 26
     constraintVer = "1.0.2"
+    crashAnalytics = "2.9.1"
     dexMakerVer = "1.2"
     espContVer = "2.2.2"
     espCoreVer = "2.2.2"
     espIntVer = "2.2.2"
     espUIAVer = "2.1.2"
-    firebaseVer = "11.0.0"
+    firebaseVer = "12.0.1"
+    flexboxVer = "0.3.1"
     glideVer = "4.0.0-RC0"
-    playServicesVer = "11.0.0"
+    playServicesVer = firebaseVer
     hamcrestVer = "1.3"
     jacocoVer = "0.7.9"
     junitVer = "4.12"
+    lapismSearchVer = "5.0.0-alpha7"
     minSdkVer = 21
     mockitoVer = "2.9.0"
     roboElecVer = "3.4.2"
     rxJavaVer = "1.3.0"
-    supportVer = "25.4.0"
-    targetSdkVer = 25
-    verCode = 1
-    verName = "1.0"
+    rxJava2Ver = "2.1.6"
+    supportVer = "27.1.0"
+    targetSdkVer = 26
+    verCode = 9
+    verName = "1.7-beta"
 }
 
 def configDir = "${project.rootDir}/config/"
