@@ -1,20 +1,25 @@
 package com.macbitsgoa.ard;
 
 import android.app.Application;
-import android.content.Context;
 
 import com.google.firebase.database.FirebaseDatabase;
-
 import com.macbitsgoa.ard.utils.AHC;
+
 import io.realm.Realm;
 import io.realm.RealmConfiguration;
 
 /**
  * Custom class extending the Application class.
  * This is used for setting up libraries.
+ *
  * @author Vikramaditya Kukreja
  */
 public class ARD extends Application {
+
+    /**
+     * TAG for the class.
+     */
+    public static final String TAG = ARD.class.getSimpleName();
 
     @Override
     public void onCreate() {
@@ -24,12 +29,8 @@ public class ARD extends Application {
         final RealmConfiguration realmConfiguration = new RealmConfiguration.Builder()
                 .name(AHC.REALM_ARD_DATABASE)
                 .schemaVersion(AHC.REALM_ARD_DATABASE_SCHEMA)
+                .deleteRealmIfMigrationNeeded()
                 .build();
         Realm.setDefaultConfiguration(realmConfiguration);
-    }
-
-    @Override
-    protected void attachBaseContext(final Context base) {
-        super.attachBaseContext(base);
     }
 }
