@@ -6,6 +6,8 @@ import android.support.v4.app.Fragment;
 import android.support.v4.app.FragmentManager;
 import android.support.v4.app.FragmentPagerAdapter;
 
+import com.macbitsgoa.ard.fragments.forum.GeneralFragment;
+
 import java.util.ArrayList;
 import java.util.List;
 
@@ -19,12 +21,12 @@ public class ViewPagerAdapter extends FragmentPagerAdapter {
     /**
      * List to hold fragments objects.
      */
-    private List<Fragment> fragmentList = new ArrayList<>();
+    private final List<Fragment> fragmentList = new ArrayList<>();
 
     /**
      * List to hold fragment titles.
      */
-    private List<String> fragmentTitlesList = new ArrayList<>();
+    private final List<String> fragmentTitlesList = new ArrayList<>();
 
     /**
      * Constructor with args.
@@ -33,6 +35,10 @@ public class ViewPagerAdapter extends FragmentPagerAdapter {
      */
     public ViewPagerAdapter(@NonNull final FragmentManager manager) {
         super(manager);
+        addFragment(GeneralFragment.newInstance("0"), "Ph.D.");
+        addFragment(GeneralFragment.newInstance("1"), "M.E.");
+        addFragment(GeneralFragment.newInstance("2"), "B.E.");
+        addFragment(GeneralFragment.newInstance("3"), "Others");
     }
 
     @Override
