@@ -1,103 +1,94 @@
 package com.macbitsgoa.ard.fragments;
 
-import android.animation.Animator;
 import android.content.Intent;
+import android.content.res.Configuration;
 import android.os.Bundle;
+import android.os.Handler;
 import android.support.annotation.NonNull;
-import android.support.annotation.Nullable;
-import android.support.constraint.ConstraintLayout;
-import android.support.design.widget.FloatingActionButton;
-import android.support.v4.app.ActivityOptionsCompat;
+import android.support.design.widget.AppBarLayout;
+import android.support.design.widget.CoordinatorLayout;
 import android.support.v4.app.Fragment;
-import android.support.v7.widget.CardView;
+import android.support.v4.view.ViewPager;
+import android.support.v4.widget.NestedScrollView;
+import android.support.v7.widget.DividerItemDecoration;
 import android.support.v7.widget.LinearLayoutManager;
 import android.support.v7.widget.RecyclerView;
 import android.view.LayoutInflater;
 import android.view.View;
-import android.view.ViewAnimationUtils;
 import android.view.ViewGroup;
-import android.view.animation.AnimationUtils;
 
-import com.google.firebase.database.DataSnapshot;
-import com.google.firebase.database.DatabaseError;
-import com.google.firebase.database.DatabaseReference;
-import com.google.firebase.database.FirebaseDatabase;
-import com.google.firebase.database.ValueEventListener;
-import com.macbitsgoa.ard.BuildConfig;
 import com.macbitsgoa.ard.R;
-import com.macbitsgoa.ard.activities.PostActivity;
-import com.macbitsgoa.ard.activities.PostDetailsActivity;
-import com.macbitsgoa.ard.activities.SearchActivity;
+import com.macbitsgoa.ard.activities.AnnActivity;
+import com.macbitsgoa.ard.adapters.AnnSlideshowAdapter;
 import com.macbitsgoa.ard.adapters.HomeAdapter;
-import com.macbitsgoa.ard.interfaces.HomeFragmentListener;
-import com.macbitsgoa.ard.interfaces.OnItemClickListener;
-import com.macbitsgoa.ard.interfaces.RecyclerItemClickListener;
-import com.macbitsgoa.ard.keys.AnnItemKeys;
-import com.macbitsgoa.ard.keys.PostKeys;
-import com.macbitsgoa.ard.models.AnnItem;
-import com.macbitsgoa.ard.models.TypeItem;
-import com.macbitsgoa.ard.types.PostType;
-import com.macbitsgoa.ard.utils.AHC;
-
-import java.util.ArrayList;
-import java.util.List;
+import com.macbitsgoa.ard.adapters.SlideshowAdapter;
 
 import butterknife.BindView;
 import butterknife.ButterKnife;
+import butterknife.OnClick;
 import butterknife.Unbinder;
-import io.realm.Realm;
-import io.realm.Sort;
 
 /**
- * A simple {@link Fragment} subclass.
- * Activities that contain this fragment can implement the
- * {@link HomeFragmentListener} interface
- * to handle interaction events.
- * Use the {@link HomeFragment#newInstance} factory method to
- * create an instance of this fragment.
+ * A simple {@link Fragment} subclass to display home content.
  *
  * @author Vikramaditya Kukreja
  */
-public class HomeFragment extends Fragment implements View.OnClickListener, OnItemClickListener {
+public class HomeFragment extends BaseFragment implements AppBarLayout.OnOffsetChangedListener {
+
+    /**
+     * TAG for this class.
+     */
+    public static final String TAG = HomeFragment.class.getSimpleName();
 
     /**
      * RecyclerView to display Home content.
      */
     @BindView(R.id.recyclerView_fragment_home)
-    public RecyclerView recyclerView;
+    RecyclerView homeRV;
 
     /**
-     * Main FAB for fragment. Has multiple sub mini FABs.
+     * ViewPager for image slideshow.
      */
-    @BindView(R.id.fab_fragment_home_add)
-    public FloatingActionButton mainFab;
+    @BindView(R.id.vp_fragment_home_slideshow)
+    ViewPager slideshowVP;
 
     /**
-     * Sub mini FAB for general upload.
+     * Top appbar layout to hide in case of orientation change.
      */
-    @BindView(R.id.fab_fragment_home_announce)
-    public FloatingActionButton announceFab;
-
-    @BindView(R.id.cardView_frame_search)
-    public CardView cvfs;
+    @BindView(R.id.ab_fragment_home)
+    AppBarLayout appBarLayout;
 
     /**
-     * A simple {@code View} object which has a custom background to be used when main FAB is
-     * clicked.
+     * Viewpager to show ann data.
      */
-    @BindView(R.id.view_fragment_home_backdrop)
-    public View backdrop;
+    @BindView(R.id.vp_vh_announcement)
+    ViewPager annVP;
 
     /**
-     * Status boolean to maintain current status of {@link HomeFragment#mainFab}
-     * in {@link HomeFragment}.
+     * Nested scroll view for ann and rv.
      */
-    public boolean isFabOpen;
+    @BindView(R.id.nsv_fragment_home)
+    NestedScrollView nsv;
 
     /**
-     * HomeAdapter object.
+     * Handler to run {@link #imageSlideshowRunnable}.
      */
-    public HomeAdapter homeAdapter;
+    private Handler imageSlideshowHandler;
+
+    /**
+     * Runnable for images slideshow.
+     */
+    private Runnable imageSlideshowRunnable;
+
+    /**
+     * Handler to run {@link #annSlideshowRunable}.
+     */
+    private Handler annSlideshowHandler;
+
+    /**
+     * Runnable for ann slideshow text.
+     */
+    private Runnable annSlideshowRunable;
 
     /**
      * Unbinder for ButterKnife.
@@ -105,300 +96,134 @@ public class HomeFragment extends Fragment implements View.OnClickListener, OnIt
     private Unbinder unbinder;
 
     /**
-     * Reference to node {@link AHC#FDR_HOME} to which listener is attached.
+     * Slideshow adapter.
      */
-    private DatabaseReference dbRef = FirebaseDatabase.getInstance()
-            .getReference().child(BuildConfig.BUILD_TYPE).child(AHC.FDR_HOME);
+    private SlideshowAdapter slideshowAdapter;
 
     /**
-     * Handle for Realm instance.
+     * Announcement slideshow adapter.
      */
-    private Realm database;
+    private AnnSlideshowAdapter annAdapter;
 
     /**
-     * EventListener for {@link AHC#FDR_HOME} which is required to remove in {@link #onStop()}.
+     * {@link View#offsetTopAndBottom(int)} of {@link #appBarLayout}.
      */
-    private ValueEventListener homeEventListener;
-
-    /**
-     * Item touch listener of RecyclerView.
-     */
-    private RecyclerView.OnItemTouchListener onItemTouchListener;
-
-    /**
-     * Use this factory method to create a new instance of
-     * this fragment using the provided parameters.
-     *
-     * @param bundle Bundled data for this fragment.
-     * @return A new instance of fragment HomeFragment.
-     */
-    public static HomeFragment newInstance(@Nullable final Bundle bundle) {
-        final HomeFragment fragment = new HomeFragment();
-        fragment.setArguments(bundle);
-        return fragment;
-    }
+    private int appBarOffset;
 
     @Override
-    public View onCreateView(final LayoutInflater inflater, final ViewGroup container,
+    public View onCreateView(@NonNull final LayoutInflater inflater, final ViewGroup container,
                              final Bundle savedInstanceState) {
         // Inflate the layout for this fragment
         final View view = inflater.inflate(R.layout.fragment_home, container, false);
-
         unbinder = ButterKnife.bind(this, view);
-
-        init();
+        homeRV.setHasFixedSize(true);
+        homeRV.setLayoutManager(new LinearLayoutManager(getContext()));
+        homeRV.addItemDecoration(new DividerItemDecoration(getContext(), DividerItemDecoration.VERTICAL));
+        homeRV.setAdapter(new HomeAdapter(getContext()));
+        scrollToTop();
+        setupSlideshows();
 
         return view;
-    }
-
-    /**
-     * View updates and listeners can be done here.
-     */
-    private void init() {
-        backdrop.setVisibility(View.INVISIBLE);
-        announceFab.setVisibility(View.INVISIBLE);
-
-        homeAdapter = new HomeAdapter();
-
-        recyclerView.setHasFixedSize(true);
-        recyclerView.setLayoutManager(new LinearLayoutManager(getContext()));
-        recyclerView.setAdapter(homeAdapter);
-
-        onItemTouchListener = new RecyclerItemClickListener(getContext(), recyclerView, this);
-        recyclerView.addOnItemTouchListener(onItemTouchListener);
-        mainFab.setOnClickListener(this);
-        announceFab.setOnClickListener(this);
-        backdrop.setOnClickListener(this);
-        cvfs.setOnClickListener(this);
     }
 
     @Override
     public void onStart() {
         super.onStart();
-        database = Realm.getDefaultInstance();
-
-        setupData(generateList());
-
-        homeEventListener = getValueEventListener();
-        dbRef.addValueEventListener(homeEventListener);
-    }
-
-    /**
-     * Method to set realm database.
-     * If argument is null, default realm database will be used.
-     * This method is convenient while testing.
-     *
-     * @param list List of {@link TypeItem} to be used in adapter.
-     */
-    //@VisibleForTesting
-    public void setupData(@NonNull final List<TypeItem> list) {
-        homeAdapter.getData().clear();
-        homeAdapter.getData().addAll(list);
-        homeAdapter.notifyDataSetChanged();
-    }
-
-    /**
-     * When fragment is created the list is generated from the realm database.
-     * This cannot be called before {@link #onStart()} as database is not ready.
-     *
-     * @return Generated list from Realm.
-     */
-    private List<TypeItem> generateList() {
-        //Clear existing results
-        final List<TypeItem> list = new ArrayList<>();
-
-        //Generate Announcement type list
-        for (final AnnItem annItem
-                : database.where(AnnItem.class).findAllSorted("date", Sort.DESCENDING)) {
-            list.add(new TypeItem(annItem, PostType.ANNOUNCEMENT));
-        }
-        return list;
-    }
-
-    /**
-     * Returns the valueEventListener for the node {@link AHC#FDR_HOME}.
-     *
-     * @return valueEventListener object that can be later detached in {@link #onStop()}.
-     */
-    @NonNull
-    private ValueEventListener getValueEventListener() {
-        return new ValueEventListener() {
-            @Override
-            public void onDataChange(final DataSnapshot dataSnapshot) {
-                for (final DataSnapshot childSnapshot : dataSnapshot.getChildren()) {
-                    final AnnItem annItem = childSnapshot.getValue(AnnItem.class);
-                    annItem.setKey(childSnapshot.getKey());
-                    database.beginTransaction();
-                    database.copyToRealmOrUpdate(annItem);
-                    database.commitTransaction();
-                }
-                setupData(generateList());
-            }
-
-            @Override
-            public void onCancelled(final DatabaseError databaseError) {
-                throw databaseError.toException();
-            }
-        };
-    }
-
-    @Override
-    public void onStop() {
-        super.onStop();
-        database.close();
-        dbRef.removeEventListener(homeEventListener);
+        appBarLayout.addOnOffsetChangedListener(this);
+        //hide app bar if orientation is landscape on starting
+        hideAppBar();
+        appBarLayout.offsetTopAndBottom(appBarOffset);
     }
 
     @Override
     public void onDestroyView() {
         super.onDestroyView();
-        recyclerView.removeOnItemTouchListener(onItemTouchListener);
-        homeAdapter = null;
+        imageSlideshowHandler.removeCallbacks(imageSlideshowRunnable);
+        annSlideshowHandler.removeCallbacks(annSlideshowRunable);
+        annAdapter.notifyOfDesctruction();
+        slideshowAdapter.notifyOfDestruction();
         unbinder.unbind();
     }
 
     @Override
-    public void onClick(final View v) {
-        final int id = v.getId();
-        if (id == R.id.fab_fragment_home_add || id == R.id.view_fragment_home_backdrop) {
-            animateFab();
-        } else if (id == R.id.cardView_frame_search) {
-            ActivityOptionsCompat options = ActivityOptionsCompat.
-                    makeSceneTransitionAnimation(getActivity(), cvfs, "profile");
-            startActivity(new Intent(getContext(), SearchActivity.class), options.toBundle());
-        } else {
-            animateFab();
-            final Intent intent = new Intent(getContext(), PostActivity.class);
-            startActivity(intent);
-        }
+    public void onOffsetChanged(final AppBarLayout appBarLayout, final int verticalOffset) {
+        appBarOffset = verticalOffset;
     }
 
     /**
-     * Method to animate main fab and invisible ones.
+     * Method to setup both slideshows.
      */
-    public void animateFab() {
-        final Animator startAnimator = ViewAnimationUtils.createCircularReveal(backdrop,
-                (int) mainFab.getX() + mainFab.getWidth() / 2,
-                (int) mainFab.getY() + mainFab.getHeight() / 2,
-                0,
-                (float) Math.hypot(backdrop.getHeight(), backdrop.getWidth()));
-        final Animator endAnimator = ViewAnimationUtils.createCircularReveal(backdrop,
-                (int) mainFab.getX() + mainFab.getWidth() / 2,
-                (int) mainFab.getY() + mainFab.getHeight() / 2,
-                (float) Math.hypot(backdrop.getHeight(), backdrop.getWidth()),
-                0);
-        startAnimator.setDuration((long) AHC.ANIMATION_MULTIPLIER * getResources()
-                .getInteger(R.integer.anim_fab_duration));
-        endAnimator.setDuration((long) AHC.ANIMATION_MULTIPLIER * getResources()
-                .getInteger(R.integer.anim_fab_duration));
+    private void setupSlideshows() {
+        //Image slideshow
+        slideshowAdapter = new SlideshowAdapter();
+        //Ann slideshow
+        annAdapter = new AnnSlideshowAdapter();
 
-        if (isFabOpen) {
-            startAnimator.removeAllListeners();
-            endAnimator.addListener(getEndAnimatorListener());
-            endAnimator.start();
-        } else {
-            endAnimator.removeAllListeners();
-            startAnimator.addListener(getStartAnimatorListener());
-            startAnimator.start();
-        }
-    }
+        imageSlideshowHandler = new Handler();
+        annSlideshowHandler = new Handler();
 
-    /**
-     * Returns listener for start animation of backdrop.
-     *
-     * @return Animator listener for animator object.
-     */
-    private Animator.AnimatorListener getStartAnimatorListener() {
-        return new Animator.AnimatorListener() {
-            @Override
-            public void onAnimationStart(final Animator animation) {
-                mainFab.startAnimation(AnimationUtils.loadAnimation(getContext(),
-                        R.anim.rotate_clock));
-                announceFab.startAnimation(AnimationUtils.loadAnimation(getContext(),
-                        R.anim.fab_open));
-                backdrop.setVisibility(View.VISIBLE);
-                announceFab.setVisibility(View.VISIBLE);
-            }
-
-            @Override
-            public void onAnimationEnd(final Animator animation) {
-                announceFab.setClickable(true);
-                backdrop.setClickable(true);
-                isFabOpen = true;
-
-                //To keep coverage at 100%
-                onAnimationRepeat(animation);
-                onAnimationCancel(animation);
-            }
-
-            @Override
-            public void onAnimationCancel(final Animator animation) {
-                //nothing is cancelled.
-            }
-
-            @Override
-            public void onAnimationRepeat(final Animator animation) {
-                //No repeating set.
-            }
+        imageSlideshowRunnable = () -> {
+            int newPos = slideshowVP.getCurrentItem() + 1;
+            newPos %= slideshowAdapter.getCount();
+            slideshowVP.setCurrentItem(newPos, true);
+            imageSlideshowHandler.postDelayed(imageSlideshowRunnable,
+                    getInteger(R.integer.image_slideshow_period));
         };
+        annSlideshowRunable = () -> {
+            if (annAdapter.getCount() == 0) {
+                return;
+            }
+            int newPos = annVP.getCurrentItem() + 1;
+            newPos %= annAdapter.getCount();
+            annVP.setCurrentItem(newPos, true);
+            annSlideshowHandler.postDelayed(annSlideshowRunable,
+                    getInteger(R.integer.ann_slideshow_period));
+        };
+
+        slideshowVP.setAdapter(slideshowAdapter);
+        annVP.setAdapter(annAdapter);
+
+        imageSlideshowHandler.postDelayed(imageSlideshowRunnable,
+                getInteger(R.integer.image_slideshow_period));
+        annSlideshowHandler.postDelayed(annSlideshowRunable,
+                getInteger(R.integer.ann_slideshow_period));
     }
 
     /**
-     * Returns listener for end animation for backdrop.
-     *
-     * @return Animator listener for animator object.
+     * Method to scroll nsv to the top.
      */
-    private Animator.AnimatorListener getEndAnimatorListener() {
-        return new Animator.AnimatorListener() {
-            @Override
-            public void onAnimationStart(final Animator animation) {
-                announceFab.startAnimation(AnimationUtils.loadAnimation(getContext(),
-                        R.anim.fab_close));
-                mainFab.startAnimation(AnimationUtils.loadAnimation(getContext(),
-                        R.anim.rotate_anticlock));
-            }
+    private void scrollToTop() {
+        nsv.fullScroll(View.FOCUS_UP);
+        nsv.scrollTo(0, 0);
 
-            @Override
-            public void onAnimationEnd(final Animator animation) {
-                announceFab.setVisibility(View.INVISIBLE);
-                backdrop.setVisibility(View.INVISIBLE);
-                announceFab.setClickable(false);
-                backdrop.setClickable(false);
-                isFabOpen = false;
+    }
 
-                //To keep coverage at 100%
-                onAnimationRepeat(animation);
-                onAnimationCancel(animation);
-            }
-
-            @Override
-            public void onAnimationCancel(final Animator animation) {
-                //Nothing is cancelled.
-            }
-
-            @Override
-            public void onAnimationRepeat(final Animator animation) {
-                //No repeating set.
-            }
-        };
+    @OnClick(R.id.itemView_ann_card)
+    public void openAnnActivity() {
+        startActivity(new Intent(getContext(), AnnActivity.class));
     }
 
     @Override
-    public void onItemClick(final View view, final int position) {
-        final Intent intent = new Intent(getContext(), PostDetailsActivity.class);
-        final TypeItem item = homeAdapter.getData().get(position);
-        intent.putExtra(PostKeys.TYPE, item.getType());
-        //switch (item.getType()) {
-        //    case PostType.ANNOUNCEMENT: {
-        intent.putExtra(AnnItemKeys.PRIMARY_KEY, ((AnnItem) item.getData()).getKey());
-        //        break;
-        //    }
-        //}
-        startActivity(intent);
+    public void onConfigurationChanged(final Configuration newConfig) {
+        super.onConfigurationChanged(newConfig);
+        hideAppBar();
     }
 
-    @Override
-    public void onLongItemClick(final View view, final int position) {
-        //Not required as of now
+    /**
+     * Function to hide appbar depending on orientation.
+     */
+    private void hideAppBar() {
+        final CoordinatorLayout.LayoutParams params
+                = (CoordinatorLayout.LayoutParams) nsv.getLayoutParams();
+
+        // Checks the orientation of the screen
+        if (getResources().getConfiguration().orientation == Configuration.ORIENTATION_LANDSCAPE) {
+            params.setBehavior(null);
+            appBarLayout.setVisibility(View.GONE);
+        } else if (getResources().getConfiguration().orientation == Configuration.ORIENTATION_PORTRAIT) {
+            params.setBehavior(new AppBarLayout.ScrollingViewBehavior());
+            appBarLayout.setVisibility(View.VISIBLE);
+        }
+        nsv.requestLayout();
     }
 }
