@@ -11,26 +11,21 @@ public class AnnItemKeys {
      * Key for <b>primary key</b> field in Realm Database.
      * {@link com.macbitsgoa.ard.models.AnnItem#key}
      */
-    public static final String PRIMARY_KEY = "key";
-
-    /**
-     * Key for {@link com.macbitsgoa.ard.types.PostType#ANNOUNCEMENT} field.
-     */
-    public static final String TYPE_KEY = "type";
+    public static final String KEY = "key";
 
     /**
      * Key for {@link com.macbitsgoa.ard.models.AnnItem#data} field.
      */
-    public static final String DATA_KEY = "data";
+    public static final String DATA = "data";
 
     /**
      * Key for {@link com.macbitsgoa.ard.models.AnnItem#author} field.
      */
-    public static final String AUTHOR_KEY = "author";
+    public static final String AUTHOR = "author";
 
     /**
      * Key for {@link com.macbitsgoa.ard.models.AnnItem#date} field.
      */
-    public static final String DATE_KEY = "date";
+    public static final String DATE = "date";
 
 }
