@@ -1,5 +1,7 @@
 package com.macbitsgoa.ard.models;
 
+import android.support.annotation.NonNull;
+
 /**
  * Item format that can store any data and a type value.
  *
@@ -10,13 +12,13 @@ public class TypeItem {
     /**
      * Data for the object depending on {@link #type}.
      */
-    private Object data;
+    private final Object data;
 
     /**
      * An int value corresponding to the type of data.
      * Eg. {@link com.macbitsgoa.ard.types.PostType#ANNOUNCEMENT}
      */
-    private int type;
+    private final int type;
 
     /**
      * Constructor for type item.
@@ -24,7 +26,7 @@ public class TypeItem {
      * @param data Object data.
      * @param type int value of type declared in an util class.
      */
-    public TypeItem(final Object data, final int type) {
+    public TypeItem(@NonNull final Object data, final int type) {
         this.data = data;
         this.type = type;
     }
