@@ -2,6 +2,7 @@ package com.macbitsgoa.ard.types;
 
 /**
  * Utility class that has useful constants for PostActivity.
+ *
  * @author Vikramaditya Kukreja
  */
 public class PostType {
@@ -9,5 +10,5 @@ public class PostType {
     /**
      * Type for announcement posts.
      */
-    public static final int ANNOUNCEMENT = 0;
+    public static final int ANNOUNCEMENT = -64;
 }
