@@ -1,10 +1,14 @@
 package com.macbitsgoa.ard.viewholders;
 
 import android.support.v7.widget.RecyclerView;
+import android.text.method.LinkMovementMethod;
 import android.view.View;
 import android.widget.TextView;
 
 import com.macbitsgoa.ard.R;
+
+import butterknife.BindView;
+import butterknife.ButterKnife;
 
 /**
  * ViewHolder for {@link com.macbitsgoa.ard.types.PostType#ANNOUNCEMENT}.
@@ -14,13 +18,20 @@ import com.macbitsgoa.ard.R;
 public class AnnViewHolder extends RecyclerView.ViewHolder {
 
     /**
+     * TAG for class.
+     */
+    public static final String TAG = AnnViewHolder.class.getSimpleName();
+
+    /**
      * Author of the post.
      */
+    @BindView(R.id.textView_viewHolder_ann_extras)
     public TextView extras;
 
     /**
      * Announcement data.
      */
+    @BindView(R.id.textView_viewHolder_ann_data)
     public TextView data;
 
     /**
@@ -30,7 +41,8 @@ public class AnnViewHolder extends RecyclerView.ViewHolder {
      */
     public AnnViewHolder(final View itemView) {
         super(itemView);
-        data = (TextView) itemView.findViewById(R.id.textView_viewHolder_ann_data);
-        extras = (TextView) itemView.findViewById(R.id.textView_viewHolder_ann_extras);
+        ButterKnife.bind(this, itemView);
+        data.setMovementMethod(LinkMovementMethod.getInstance());
+        extras.setMovementMethod(LinkMovementMethod.getInstance());
     }
 }
